Custom styling and transitioning
Custom classes were added to the global navigation tray to apply the similar styles to what is natively applied without having to rely on the randomly generated class names from Canvas.  Cloning could potentially be used to replace this in the future, but it at least has it load initially with a similar look and feel since cloning isn't always available before a custom tray is first opened.  *Note: No changes to the responsive menu nor loading icon for callback function with this update.

Simplified HTML for close button while still applying similar styles.

Focus on the close button when the tray is opened.

Transition effect applied when opening tray and when closing tray with close button or click outside of tray.  The tray starts off-screen and the gnct_open class is applied to move it on-screen.  TO slide out, the gnct_open class is removed.